Recomputing safety area border and obstacles with a change of world origin
In scenarios where a safety border and obstacles are set to respect to
the world origin, if a change of world origin is done the safety area
stays in the global coordinate frame (latlon) position, and doesn't
follow the new world origin. Given that the safety area was set to
respect to the previous world origin.